+++ NEWS.md
@@ -3,6 +3,7 @@
 ## BUG FIXES
 
 * get_cid() became more robust to smiles queries with special characters.
+* `bcpc_query` now correctly parses the activity in cases that two activities are given (e.g. "herbicides" and "plant growth regulators")
 
 # webchem 1.1.2